tachyon: adding a mechanism for storing initial objects; cosmodrone: preparation for dynamic objects

+++ engine/tachyon_mesh_manager.h
@@ -14,7 +14,7 @@ tMesh Tachyon_LoadMesh(const char* path, const tVec3f& axis_factors = tVec3f(1.f
 tMesh Tachyon_CreatePlaneMesh();
 tMesh Tachyon_CreateCubeMesh();
 tMesh Tachyon_CreateSphereMesh(uint8 divisions);
-uint32 Tachyon_AddMesh(Tachyon* tachyon, const tMesh& mesh, uint16 total);
+uint16 Tachyon_AddMesh(Tachyon* tachyon, const tMesh& mesh, uint16 total);
 void Tachyon_InitializeObjects(Tachyon* tachyon);
 tObject& Tachyon_CreateObject(Tachyon* tachyon, uint16 mesh_index);
 void Tachyon_RemoveObject(Tachyon* tachyon, tObject& object);
